Moved logo and statue of liberty images to styles folder & adjusted position of statue
