+++ src/styles/statueOfLiberty.jpg
